Updated Storyboard for Issue 7
Based on changes made on Monday's Google meet with online group

For each chart, I'm not exactly sure how to make the results look more striking if that makes sense. A lot of the charts come out as "samey" and hard to differntiate, but I'm not sure how to do that without changing the data around too much or using a chart that won't exactly work in demonstration.